update: remove tvl information about options vaults
